debug prestamos

+++ src/prestamos.hpp
@@ -30,27 +30,27 @@ class Prestamos{
         std::vector<int> mesesHipotecario = {120, 180, 240};
 
         //Estos son los intereses anuales para los prestamos personales
-        int interesPersonalAnualColones = 0.2;
-        int interesPersonalAnualDolar = 0.16;
+        float interesPersonalAnualColones = 0.2;
+        float interesPersonalAnualDolar = 0.16;
 
         //Estos son los intereses anuales para los prestamos prendarios
-        int interesPrendarioAnualColones = 0.155;
-        int interesPrendarioAnualDolar = 0.125;
+        float interesPrendarioAnualColones = 0.155;
+        float interesPrendarioAnualDolar = 0.125;
 
         //Estos son los intereses anuales para los prestmos hipotecarios
-        int interesHipotecarioAnualColones = 0.12;
-        int interesHipotecarioAnualDolar = 0.07;
+        float interesHipotecarioAnualColones = 0.12;
+        float interesHipotecarioAnualDolar = 0.07;
 
-        int tasaCompraDolarColones = 521.5; //Asumiendo que la gente compra dolar para pagar el prestamo.
+        float tasaCompraDolarColones = 521.5; //Asumiendo que la gente compra dolar para pagar el prestamo.
 
         //Funcion que cambia el interes anual a mensual, necesario para dar las cuotas mensuales
-        int interesAnualaMensual(int interes);
+        float interesAnualaMensual(float interes);
 
         //Calcula las cuotas utilizadas para la tabla personalizada
-        std::vector<int> calcularCoutas(int interes ,std::vector<int> meses, double monto, double monto_prendario = 0);
+        std::vector<int> calcularCoutas(float interes ,std::vector<int> meses, double monto, double monto_prendario = 0);
 
         //Imprime la informacion en una tabla para poder desplegar las valores de las cuotas
-        void imprimirTablaInformacion(int interesColon, int interesDolar, std::vector<int> cuotas_dolar,std::vector<int> cuotas_colon, std::vector<int> meses ); 
+        void imprimirTablaInformacion(float interesColon, float interesDolar, std::vector<int> cuotas_dolar,std::vector<int> cuotas_colon, std::vector<int> meses ); 
 
         //Valida el si la persona es apta para el prestamo.
         bool validacionPrestamo(std::vector<int> meses, std::vector<int> cuotas_dolar, std::vector<int> cuotas_colon, double salario, int tipoMoneda);
